must be logged into associated account to CRUD listitems

+++ src/routes/list.js
@@ -2,7 +2,7 @@
 
 const express = require('express');
 const db = require('../models/index.js');
-// const authenticateBearer = require('../auth/middleware/bearer')
+const authenticateBearer = require('../auth/middleware/bearer')
 
 // const data = require('../models/index.js');
 // const { QueryTypes } = require('sequelize');
@@ -10,19 +10,12 @@ const router = express.Router();
 
 
 // === === routers === === //
-// --- comment these back in when bearer/auth is functional --- //
-// router.post('/lists', authenticateBearer, create)
-// router.get('/lists', getAll);
-// router.get('/lists/:listsId', getOne);
-// router.put('/lists/:listsId', authenticateBearer, update);
-// router.delete('/lists/:listsId', authenticateBearer, remove);
-
-router.post('/listitem', create)
-router.get('/listitem', getAll);
-router.get('/listitem/:itemId', getOne);
-router.put('/listitem/:itemId', update);
-router.delete('/listitem/:itemId', remove);
-
+router.post('/listitem', authenticateBearer, create)
+router.get('/listitem', authenticateBearer, getAll);
+router.get('/listitem/:itemId', authenticateBearer, getOne);
+router.put('/listitem/:itemId', authenticateBearer, update);
+router.delete('/listitem/:itemId', authenticateBearer, remove);
+// === === === === === === //
 
 async function create(req, res) {
   // body expects: name, quantity, category, price, notes, image
@@ -30,23 +23,30 @@ async function create(req, res) {
   // console.log('request user 🥩', request.user)
   const { productName, quantity, category, price, notes, image } = req.body
   console.log({ productName, quantity, category, price, notes, image })
-  const newList = await db.list.create({ userId: 1, productName, quantity, category, price, notes, image })
+  const newList = await db.list.create({ userId: req.user.dataValues.id, productName, quantity, category, price, notes, image })
   console.log('new list: ', newList)
 
   res.status(200).send({ success: true, data: newList, message: true ? 'Created!' : 'Error Creating!' });
 }
 
+// const servicesObject = request.body;
+// servicesObject.freelancer = request.user.dataValues.id
+// const servicesData = await data.services.create(servicesObject);
+// console.log('SERVICES DATA 🥐', servicesData)
+
 async function getAll(req, res) {
-  console.log(db.list)
-  const allLists = await db.list.findAll();
-  console.log(allLists)
+  // console.log(db.list)
+  console.log('logged in user ID is: ', req.user.dataValues.id)
+  const allLists = await db.list.findAll({ where: { userId: req.user.dataValues.id } });
+  // console.log(allLists)
 
   res.status(200).send({ success: true, data: allLists })
 }
 
 async function getOne(req, res) {
   const { itemId } = req.params
-  const singleItem = await db.list.findOne({ id: itemId });
+  console.log('REQ DOT USER: ', req.user)
+  const singleItem = await db.list.findOne({ where: { id: itemId, userId: req.user.dataValues.id } });
 
   res.status(200).send({ success: true, data: singleItem })
 }
@@ -54,12 +54,12 @@ async function getOne(req, res) {
 async function update(req, res) {
   const { itemId } = req.params
   const { productName, quantity, category, price, notes, image } = req.body
-  const oneItem = await db.list.findOne({ where: { id: itemId } });
+  const oneItem = await db.list.findOne({ where: { id: itemId, userId: req.user.dataValues.id } });
   if (!oneItem) {
     return res.status(409).send({ success: false, data: [], message: 'Item Doesn\'t Exist' })
   }
 
-  const updatedList = await oneItem.update({ productName, quantity, category, price, notes, image }, { where: { id: itemId } })
+  const updatedList = await oneItem.update({ productName, quantity, category, price, notes, image }, { where: { id: itemId, userId: req.user.dataValues.id } })
   console.log('updated list: ', updatedList)
 
   res.status(200).send({ success: true, data: updatedList })
@@ -67,7 +67,7 @@ async function update(req, res) {
 
 async function remove(req, res) {
   const { itemId } = req.params
-  const deletedItem = await db.list.destroy({ where: { id: itemId } });
+  const deletedItem = await db.list.destroy({ where: { id: itemId, userId: req.user.dataValues.id } });
 
   res.status(200).send({ success: true, data: deletedItem })
 }
